Fix verbose_display freeze
If you forgot to place you images parameter in a list
when calling verbose_display, your computer would freeze
because it would begin parsing every line of the image
as if it were an image itself.
Now there is a check to ensure it is in a list, and if not,
to put it in one.